+++ packages/web/src/app/[domain]/chat/components/newChatPanel.tsx
@@ -51,7 +51,6 @@ export const NewChatPanel = ({
                             languageModels={languageModels}
                             selectedSearchScopes={selectedSearchScopes}
                             searchContexts={searchContexts}
-                            onContextSelectorOpenChanged={setIsContextSelectorOpen}
                         />
                         <div className="w-full flex flex-row items-center bg-accent rounded-b-md px-2">
                             <ChatBoxToolbar